@@ -52,12 +52,15 @@
   },
   "homepage": "https://github.com/kenny-hibino/react-places-autocomplete#readme",
   "devDependencies": {
-    "babel-cli": "6.10.1",
-    "babel-loader": "6.2.4",
-    "babel-preset-es2015": "6.9.0",
-    "babel-preset-react": "6.11.1",
-    "babel-preset-stage-2": "6.11.0",
-    "babel-register": "6.9.0",
+    "babel-cli": "^6.23.0",
+    "babel-loader": "^6.3.2",
+    "babel-plugin-react-transform": "^2.0.2",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-react-remove-prop-types": "^0.3.2",
+    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-react": "^6.23.1",
+    "babel-preset-stage-0": "^6.22.0",
+    "babel-register": "^6.23.0",
     "chai": "3.5.0",
     "chai-enzyme": "0.5.0",
     "commitizen": "2.8.2",
@@ -76,7 +79,7 @@
     "rimraf": "2.5.3",
     "semantic-release": "^4.3.5",
     "sinon": "1.17.4",
-    "webpack": "1.13.1",
+    "webpack": "^1.14.0",
     "webpack-dev-server": "^1.16.2"
   },
   "config": {
