@@ -4,74 +4,313 @@
 * See https://kenny-hibino.github.io/react-places-autocomplete
 */
 
-import React from 'react'
-import PlacesAutocompleteWithTypeAhead from './PlacesAutocompleteWithTypeAhead'
-import PlacesAutocompleteBasic from './PlacesAutocompleteBasic'
-import { mobileCheck } from './helpers'
+import React, { Component, PropTypes } from 'react'
+import defaultStyles from './defaultStyles'
 
-const PlacesAutocomplete = (props) => {
-  const { typeAhead, ...rest } = props
-  // Work around for React KeyDown event issue: https://github.com/facebook/react/issues/6176
-  const isMobile = mobileCheck()
-  if (typeAhead && !isMobile) {
-    return <PlacesAutocompleteWithTypeAhead {...rest} />
-  } else {
-    return <PlacesAutocompleteBasic {...rest} />
+class PlacesAutocomplete extends Component {
+  constructor(props) {
+    super(props)
+
+    this.state = { autocompleteItems: [] }
+
+    this.autocompleteCallback = this.autocompleteCallback.bind(this)
+    this.handleInputKeyDown = this.handleInputKeyDown.bind(this)
+    this.handleInputChange = this.handleInputChange.bind(this)
+  }
+
+  componentDidMount() {
+    this.autocompleteService = new google.maps.places.AutocompleteService()
+    this.autocompleteOK = google.maps.places.PlacesServiceStatus.OK
+  }
+
+  autocompleteCallback(predictions, status) {
+    if (status != this.autocompleteOK) {
+      this.props.onError(status)
+      if (this.props.clearItemsOnError) { this.clearAutocomplete() }
+      return
+    }
+
+    // transform snake_case to camelCase
+    const formattedSuggestion = (structured_formatting) => ({
+      mainText: structured_formatting.main_text,
+      secondaryText: structured_formatting.secondary_text,
+    })
+
+    this.setState({
+      autocompleteItems: predictions.map((p, idx) => ({
+        suggestion: p.description,
+        placeId: p.place_id,
+        active: false,
+        index: idx,
+        formattedSuggestion: formattedSuggestion(p.structured_formatting),
+      }))
+    })
+  }
+
+  clearAutocomplete() {
+    this.setState({ autocompleteItems: [] })
+  }
+
+  selectAddress(address, placeId) {
+    this.clearAutocomplete()
+    this.handleSelect(address, placeId)
+  }
+
+  handleSelect(address, placeId) {
+    this.props.onSelect ? this.props.onSelect(address, placeId) : this.props.inputProps.onChange(address)
+  }
+
+  getActiveItem() {
+    return this.state.autocompleteItems.find(item => item.active)
+  }
+
+  selectActiveItemAtIndex(index) {
+    const activeName = this.state.autocompleteItems.find(item => item.index === index).suggestion
+    this.setActiveItemAtIndex(index)
+    this.props.inputProps.onChange(activeName)
+  }
+
+  handleEnterKey() {
+    if (this.state.autocompleteItems.length === 0) {
+      return
+    }
+
+    const activeItem = this.getActiveItem()
+    if (activeItem === undefined) {
+      this.handleEnterKeyWithoutActiveItem()
+    } else {
+      this.selectAddress(activeItem.suggestion, activeItem.placeId)
+    }
+  }
+
+  handleEnterKeyWithoutActiveItem() {
+    if (this.props.onEnterKeyDown) {
+      this.props.onEnterKeyDown(this.props.inputProps.value)
+      this.clearAutocomplete()
+    } else {
+      return //noop
+    }
+  }
+
+  handleDownKey() {
+    if (this.state.autocompleteItems.length === 0) {
+      return
+    }
+
+    const activeItem = this.getActiveItem()
+    if (activeItem === undefined) {
+      this.selectActiveItemAtIndex(0)
+    } else {
+      const nextIndex = (activeItem.index + 1) % this.state.autocompleteItems.length
+      this.selectActiveItemAtIndex(nextIndex)
+    }
+  }
+
+  handleUpKey() {
+    if (this.state.autocompleteItems.length === 0) {
+      return
+    }
+
+    const activeItem = this.getActiveItem()
+    if (activeItem === undefined) {
+      this.selectActiveItemAtIndex(this.state.autocompleteItems.length - 1)
+    } else {
+      let prevIndex
+      if (activeItem.index === 0) {
+        prevIndex = this.state.autocompleteItems.length - 1
+      } else {
+        prevIndex = (activeItem.index - 1) % this.state.autocompleteItems.length
+      }
+      this.selectActiveItemAtIndex(prevIndex)
+    }
+  }
+
+  handleInputKeyDown(event) {
+    switch (event.key) {
+      case 'Enter':
+        event.preventDefault()
+        this.handleEnterKey()
+        break
+      case 'ArrowDown':
+        event.preventDefault() // prevent the cursor from moving
+        this.handleDownKey()
+        break
+      case 'ArrowUp':
+        event.preventDefault() // prevent the cursor from moving
+        this.handleUpKey()
+        break
+      case 'Escape':
+        this.clearAutocomplete()
+        break
+    }
+
+    if (this.props.inputProps.onKeyDown) {
+      this.props.inputProps.onKeyDown(event)
+    }
+  }
+
+  setActiveItemAtIndex(index) {
+    this.setState({
+      autocompleteItems: this.state.autocompleteItems.map((item, idx) => {
+        if (idx === index) {
+          return { ...item, active: true }
+        } else {
+          return { ...item, active: false }
+        }
+      }),
+    })
+  }
+
+  handleInputChange(event) {
+    this.props.inputProps.onChange(event.target.value)
+    if (!event.target.value) {
+      this.clearAutocomplete()
+      return
+    }
+    this.autocompleteService.getPlacePredictions({ ...this.props.options, input: event.target.value }, this.autocompleteCallback)
+  }
+
+  handleInputOnBlur(event) {
+    this.clearAutocomplete()
+
+    if (this.props.inputProps.onBlur) {
+      this.props.inputProps.onBlur(event)
+    }
+  }
+
+  inlineStyleFor(...props) {
+    const { classNames, styles } = this.props
+    // No inline style if className is passed via props for the element.
+    if (props.some(prop => classNames.hasOwnProperty(prop))) {
+      return {}
+    }
+
+    return props.reduce((acc, prop) => {
+      return {
+        ...acc,
+        ...defaultStyles[prop],
+        ...styles[prop],
+      }
+    }, {})
+  }
+
+  classNameFor(...props) {
+    const { classNames } = this.props
+
+    return props.reduce((acc, prop) => {
+      const name = classNames[prop] || ''
+      return name ? `${acc} ${name}` : acc
+    }, '')
+  }
+
+  getInputProps() {
+    const defaultInputProps = {
+      type: "text",
+    }
+
+    return {
+      ...defaultInputProps,
+      ...this.props.inputProps,
+      onChange: (event) => {
+        this.handleInputChange(event)
+      },
+      onKeyDown: (event) => {
+        this.handleInputKeyDown(event)
+      },
+      onBlur: (event) => {
+        this.handleInputOnBlur(event)
+      },
+      style: this.inlineStyleFor('input'),
+      className: this.classNameFor('input'),
+    }
+  }
+
+  render() {
+    const { classNames, styles } = this.props
+    const { autocompleteItems } = this.state
+    const inputProps = this.getInputProps()
+
+    return (
+      <div
+        id="PlacesAutocomplete__root"
+        style={this.inlineStyleFor('root')}
+        className={this.classNameFor('root')}>
+        <input {...inputProps} />
+        {autocompleteItems.length > 0 && (
+          <div
+            id="PlacesAutocomplete__autocomplete-container"
+            style={this.inlineStyleFor('autocompleteContainer')}
+            className={this.classNameFor('autocompleteContainer')}>
+            {autocompleteItems.map((p, idx) => (
+              <div
+                key={p.placeId}
+                onMouseOver={() => this.setActiveItemAtIndex(p.index)}
+                onMouseDown={() => this.selectAddress(p.suggestion, p.placeId)}
+                style={ p.active ? this.inlineStyleFor('autocompleteItem', 'autocompleteItemActive') :this.inlineStyleFor('autocompleteItem') }
+                className={ p.active ? this.classNameFor('autocompleteItem', 'autocompleteItemActive') : this.classNameFor('autocompleteItem') }>
+                {this.props.autocompleteItem({ suggestion: p.suggestion, formattedSuggestion: p.formattedSuggestion })}
+              </div>
+            ))}
+          </div>
+        )}
+      </div>
+    )
   }
 }
 
 PlacesAutocomplete.propTypes = {
-  value: React.PropTypes.string.isRequired,
-  onChange: React.PropTypes.func.isRequired,
-  onError: React.PropTypes.func,
-  clearItemsOnError: React.PropTypes.bool,
-  onSelect: React.PropTypes.func,
-  placeholder: React.PropTypes.string,
-  autoFocus: React.PropTypes.bool,
-  inputName: React.PropTypes.string,
-  inputId: React.PropTypes.string,
-  autocompleteItem: React.PropTypes.func,
-  classNames: React.PropTypes.shape({
-    root: React.PropTypes.string,
-    input: React.PropTypes.string,
-    autocompleteContainer: React.PropTypes.string,
-    autocompleteItem: React.PropTypes.string,
-    autocompleteItemActive: React.PropTypes.string,
+  inputProps: (props, propName) => {
+    const inputProps = props[propName];
+
+    if (!inputProps.hasOwnProperty('value')) {
+      throw new Error('\'inputProps\' must have \'value\'.')
+    }
+
+    if (!inputProps.hasOwnProperty('onChange')) {
+      throw new Error('\'inputProps\' must have \'onChange\'.')
+    }
+  },
+  onError: PropTypes.func,
+  clearItemsOnError: PropTypes.bool,
+  onSelect: PropTypes.func,
+  autocompleteItem: PropTypes.func,
+  classNames: PropTypes.shape({
+    root: PropTypes.string,
+    input: PropTypes.string,
+    autocompleteContainer: PropTypes.string,
+    autocompleteItem: PropTypes.string,
+    autocompleteItemActive: PropTypes.string,
   }),
-  styles: React.PropTypes.shape({
-    root: React.PropTypes.object,
-    input: React.PropTypes.object,
-    autocompleteContainer: React.PropTypes.object,
-    autocompleteItem: React.PropTypes.object,
-    autocompleteItemActive: React.PropTypes.object
+  styles: PropTypes.shape({
+    root: PropTypes.object,
+    input: PropTypes.object,
+    autocompleteContainer: PropTypes.object,
+    autocompleteItem: PropTypes.object,
+    autocompleteItemActive: PropTypes.object
   }),
-  options: React.PropTypes.shape({
-    bounds: React.PropTypes.object,
-    componentRestrictions: React.PropTypes.object,
-    location: React.PropTypes.object,
-    offset: React.PropTypes.oneOfType([
-      React.PropTypes.number,
-      React.PropTypes.string
+  options: PropTypes.shape({
+    bounds: PropTypes.object,
+    componentRestrictions: PropTypes.object,
+    location: PropTypes.object,
+    offset: PropTypes.oneOfType([
+      PropTypes.number,
+      PropTypes.string
     ]),
-    radius: React.PropTypes.oneOfType([
-      React.PropTypes.number,
-      React.PropTypes.string
+    radius: PropTypes.oneOfType([
+      PropTypes.number,
+      PropTypes.string
     ]),
-    types: React.PropTypes.array
+    types: PropTypes.array
   }),
-  typeAhead: React.PropTypes.bool,
 }
 
 PlacesAutocomplete.defaultProps = {
   clearItemsOnError: false,
   onError: (status) => console.error('[react-places-autocomplete]: error happened when fetching data from Google Maps API.\nPlease check the docs here (https://developers.google.com/maps/documentation/javascript/places#place_details_responses)\nStatus: ', status),
-  placeholder: 'Address',
-  autoFocus: false,
   classNames: {},
   autocompleteItem: ({ suggestion }) => (<div>{suggestion}</div>),
   styles: {},
   options: {},
-  typeAhead: false,
 }
 
 export default PlacesAutocomplete
